Refactor: 강의 상세 관련 파일들 폴더 위치 변경
- 강의소개, 목차, QnA, 수강평, 공지사항 파일 위치 변경
- Pages -> Components 로 이동
