+++ src/components/courseDetail/courseQnA/CourseQnA.jsx
