+++ src/router/AppRouter.jsx
@@ -1,9 +1,6 @@
 import { PATH } from "@router/Constants";
 import App from "../App";
-import {
-  RouterProvider,
-  createBrowserRouter,
-} from "react-router-dom";
+import { RouterProvider, createBrowserRouter } from "react-router-dom";
 
 import HomePage from "@pages/Home/HomePage";
 import NotFoundPage from "@pages/NotFound/NotFoundPage";
@@ -12,12 +9,13 @@ import CommunityPage from "@pages/Community/CommunityPage";
 import Regist from "@pages/CommunityBoard/Regist/Regist";
 import CommunityDetail from "@pages/CommunityDetail/CommunityDetail";
 
-import CourseIntroduction from "@pages/CourseDetail/CourseIntroduction";
-import CourseCurriculum from "@pages/CourseDetail/CourseCurriculum";
-import CourseQnA from "@pages/CourseDetail/CourseQnA";
-import CourseReview from "@pages/CourseDetail/CourseReview";
-import CourseNotice from "@pages/CourseDetail/CourseNotice";
+import CourseIntroduction from "@components/courseDetail/courseIntroduction/CourseIntroduction";
+import CourseCurriculum from "@components/courseDetail/courseCurriculum/CourseCurriculum";
+import CourseQnA from "@components/courseDetail/courseQnA/CourseQnA";
+import CourseReview from "@components/courseDetail/courseReview/CourseReview";
+import CourseNotice from "@components/courseDetail/courseNotice/CourseNotice";
 import CourseDetailPage from "@pages/CourseDetail/CourseDetailPage";
+
 import LecturePage from "@pages/Lecture/LecturePage";
 import CoursesPage from "@pages/Courses/CoursesPage";
 
@@ -42,16 +40,11 @@ const AppRouter = () => {
           element: <CommunityPage />,
         },
         {
-          path: PATH.COMMUNITY_DETAIL(
-            `:communityType`,
-            `:id`
-          ),
+          path: PATH.COMMUNITY_DETAIL(`:communityType`, `:id`),
           element: <CommunityDetail />,
         },
         {
-          path: PATH.COMMUNITY_REGIST(
-            `:communityType`
-          ),
+          path: PATH.COMMUNITY_REGIST(`:communityType`),
           element: <Regist />,
         },
 
@@ -65,15 +58,11 @@ const AppRouter = () => {
           element: <CourseDetailPage />,
           children: [
             {
-              path: PATH.COURSE_INTRODUCTION(
-                `:courseId`
-              ),
+              path: PATH.COURSE_INTRODUCTION(`:courseId`),
               element: <CourseIntroduction />,
             },
             {
-              path: PATH.COURSE_CURRICULUM(
-                `:courseId`
-              ),
+              path: PATH.COURSE_CURRICULUM(`:courseId`),
               element: <CourseCurriculum />,
             },
             {
@@ -81,15 +70,11 @@ const AppRouter = () => {
               element: <CourseQnA />,
             },
             {
-              path: PATH.COURSE_REVIEW(
-                `:courseId`
-              ),
+              path: PATH.COURSE_REVIEW(`:courseId`),
               element: <CourseReview />,
             },
             {
-              path: PATH.COURSE_NOTICE(
-                `:courseId`
-              ),
+              path: PATH.COURSE_NOTICE(`:courseId`),
               element: <CourseNotice />,
             },
           ],
